move js files into views, components and layout folders
