+++ web/src/components/StatsTable/SearchBar.js
