+++ web/src/views/Game/Points.js
